# Map.cs
namespace robot_controller_api;

public class Map
{
    /// Implement <see cref=""> here following the task sheet requirements
    public int Id { get; set; }
    public int Columns { get; set; }
    public int Rows { get; set; }
    public string Name { get; set; }
    public string? Description { get; set; }
    public DateTime CreatedDate { get; set; }
    public DateTime ModifiedDate { get; set; }
}