# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Robot Controller API",
      "description": "New backend service that provides resources for the Moon robot simulator.",
      "contact": {
        "name": "Shrinivas Anavekar",
        "email": "s225255918@deakin.edu.au"
      }
    },
    "paths": {
      "/api/maps": {
        "get": {
          "tags": [
            "Maps"
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "Maps"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/maps/square": {
        "get": {
          "tags": [
            "Maps"
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/maps/{id}": {
        "get": {
          "tags": [
            "Maps"
          ],
          "operationId": "GetMapById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "Maps"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Maps"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/maps/{id}/{x}-{y}": {
        "get": {
          "tags": [
            "Maps"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "x",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "y",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/robot-commands": {
        "get": {
          "tags": [
            "RobotCommands"
          ],
          "summary": "Retrieves all robot commands.",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "post": {
          "tags": [
            "RobotCommands"
          ],
          "summary": "Adds a new robot command.",
          "requestBody": {
            "description": "The robot command to add",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotCommand"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotCommand"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotCommand"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/robot-commands/move": {
        "get": {
          "tags": [
            "RobotCommands"
          ],
          "summary": "Retrieves all move-related robot commands.",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/robot-commands/{id}": {
        "get": {
          "tags": [
            "RobotCommands"
          ],
          "summary": "Retrieves a specific robot command by its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the robot command",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "tags": [
            "RobotCommands"
          ],
          "summary": "Updates an existing robot command.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the robot command to update",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "description": "The updated robot command data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotCommand"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotCommand"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/RobotCommand"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "RobotCommands"
          ],
          "summary": "Deletes a robot command by its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the robot command to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Map": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "columns": {
              "type": "integer",
              "format": "int32"
            },
            "rows": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "createdDate": {
              "type": "string",
              "format": "date-time"
            },
            "modifiedDate": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "RobotCommand": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "isMoveCommand": {
              "type": "boolean"
            },
            "createdDate": {
              "type": "string",
              "format": "date-time"
            },
            "modifiedDate": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      }
    }
  }